move parse function to iup.utils
