+++ iup/utils.py
@@ -1,3 +1,5 @@
+import datetime as dt
+import re
 from typing import List
 
 import numpy as np
@@ -258,3 +260,23 @@ def count_unique_values(array: np.ndarray) -> List[int,]:
         Number of unique values in each column of the array
     """
     return [len(np.unique(array[:, i])) for i in range(array.shape[1])]
+
+
+def parse_name_and_date(str) -> dict[str, dt.date]:
+    """
+    Parse a string to get the model name and forecast date
+
+    Args:
+        str: A string that has a pattern of "'Model name'_forecast_starts_'Forecast date'"
+
+    Return:
+        A dictionary of model name and forecast date.
+    """
+    """Given the string that is in a format of 'Model name'_forecasts_starts_'Forecast date',
+    parse Model name and Forecast date. Return a two-element tuple."""
+    pattern = re.compile("(.+)_forecast_starts_(.+)")
+    result = pattern.fullmatch(str).groups()
+    model_name = result[0]
+    date = dt.datetime.strptime(result[1], "%Y-%m-%d").date()
+
+    return {"model_name": model_name, "forecast_date": date}
